backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy>=2.0.43
psycopg2-binary>=2.9.10
alembic==1.12.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic>=2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
reportlab>=4.0.7
openpyxl>=3.1.2
pandas>=2.1.4
numpy>=1.25.2
python-dateutil>=2.8.2
